+++ public/index.html
@@ -45,18 +45,18 @@
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
     <!-- Messenger Chat plugin Code -->
-    <!-- <div id="fb-root"></div> -->
+    <div id="fb-root"></div>
 
     <!-- Your Chat plugin code -->
-    <!-- <div id="fb-customer-chat" class="fb-customerchat"></div> -->
+    <div id="fb-customer-chat" class="fb-customerchat"></div>
 
-    <!-- <script>
+    <script>
       var chatbox = document.getElementById('fb-customer-chat')
       chatbox.setAttribute('page_id', '109182371987962')
       chatbox.setAttribute('attribution', 'biz_inbox')
-    </script> -->
+    </script>
     <!-- Your SDK code -->
-    <!-- <script>
+    <script>
       window.fbAsyncInit = function () {
         FB.init({
           xfbml: true,
@@ -72,6 +72,6 @@
         js.src = 'https://connect.facebook.net/en_US/sdk/xfbml.customerchat.js'
         fjs.parentNode.insertBefore(js, fjs)
       })(document, 'script', 'facebook-jssdk')
-    </script> -->
+    </script>
   </body>
 </html>
