+++ src/components/comments/index.js
@@ -1,0 +1,4 @@
+import CommentDrawer from './commentDrawer'
+import CommentItem from './commentItem'
+
+export { CommentDrawer, CommentItem }
