+++ src/components/fields/index.js
@@ -1,0 +1,7 @@
+import Controller from './controller'
+import Input from './input'
+import Select from './select'
+import RadioInput from './radio'
+import Textarea from './textarea'
+
+export { Controller, Input, Select, RadioInput, Textarea }
