+++ src/components/popup/index.js
@@ -1,0 +1,6 @@
+import Popup from './popup'
+import PopupConfirm from './popupConfirm'
+import PopupAddress from './popupAddress'
+import PopupStatus from './popupStatus'
+
+export { Popup, PopupAddress, PopupConfirm, PopupStatus }
