+++ src/components/preloader/index.js
@@ -1,0 +1,3 @@
+import PagePreloader from './pagePreloader'
+
+export { PagePreloader }
