+++ src/components/purchase/index.js
@@ -1,0 +1,11 @@
+import PurchaseItem from './purchaseItem'
+import PurchaseItemList from './purchaseItemList'
+import PurchaseProductItem from './purchaseProductItem'
+import PurchaseProductItemList from './purchaseProductItemList'
+
+export {
+  PurchaseItem,
+  PurchaseItemList,
+  PurchaseProductItem,
+  PurchaseProductItemList
+}
