+++ src/components/shipping/index.js
@@ -1,0 +1,4 @@
+import ShippingItem from './shippingItem'
+import ShippingList from './shippingList'
+
+export { ShippingItem, ShippingList }
