+++ src/components/user/index.js
@@ -1,0 +1,5 @@
+import AddressItem from './addressItem'
+import AddressList from './addressList'
+import Sidebar from './sidebar'
+
+export { AddressItem, AddressList, Sidebar }
